\name{GeisCamilli-package}
\alias{GeisCamilli-package}
\alias{GeisCamilli}
\docType{package}
\title{
\packageTitle{GeisCamilli}
}
\description{
\packageDescription{GeisCamilli}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GeisCamilli}
\packageIndices{GeisCamilli}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GeisCamilli}

Maintainer: \packageMaintainer{GeisCamilli}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
